renamed SimpleIdentityMatchingAlgorithm<T> to SimpleIdMatchingAlgorithm<T>
